requirements.txt
bleach==6.1.0
blinker==1.8.0
click==8.1.7
colorama==0.4.6
Flask==3.0.3
Flask-Login==0.6.3
itsdangerous==2.2.0
Jinja2==3.1.3
Markdown==3.6
MarkupSafe==2.1.5
mysql-connector-python==8.3.0
python-dotenv==1.0.1
six==1.16.0
webencodings==0.5.1
Werkzeug==3.0.2